[level 3] Title: 입국심사, Time: 54.91 ms, Memory: 80.6 MB -BaekjoonHub

+++ 프로그래머스/3/43238. 입국심사/입국심사.java
@@ -5,7 +5,7 @@ class Solution {
         long max = 0;
         for(int nn:times)if(max<nn)max = nn;
         long left = 0;
-        long right = max*(long)n;
+        long right = max*n;
         long answer = 0;
         while(left<=right){
             long mid = (left+right)/2;
